updated check for parallelization of samples (lines 25-29)
Using [[1]][1] in str_split cuts the string to just DNA151124LC, so there is an array of 170 copies of that. Using [[1]][2] cuts the string to 1, 2, 3...170, which is what I assume is what we want so we can make sure we're taking clone counts and normalized clone counts from corresponding files.